# internal/api/handler/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CastMember:
    properties:
      character:
        type: string
      name:
        type: string
      profile_path:
        type: string
    type: object
...
  models.MediaItem:
    properties:
      audio_codec:
        type: string
      backdrop_path:
        type: string
      bundle_status:
        type: string
      cast:
        items:
          $ref: '#/definitions/models.CastMember'
        type: array
      created_at:
        type: string
      director:
        type: string
      duration:
        description: seconds
        type: number
      episode_number:
        type: integer
      extra_posters:
        items:
          type: string
        type: array
      file_path:
        type: string
      file_size:
        type: integer
      height:
        type: integer
      id:
        type: string
      library_id:
        type: string
      media_type:
        $ref: '#/definitions/models.MediaType'
      mpd_path:
        type: string
      overview:
        type: string
      poster_path:
        type: string
      season_number:
        type: integer
      source_fingerprint:
        type: string
      source_status:
        type: string
      sub_jobs:
        items:
          $ref: '#/definitions/models.TranscodeSubJob'
        type: array
      title:
        type: string
      tmdb_id:
        description: Enriched metadata (nullable)
        type: integer
      transcode_progress:
        type: number
      transcode_sizes:
        $ref: '#/definitions/models.TranscodeSizesInfo'
      transcode_status:
        allOf:
        - $ref: '#/definitions/models.TranscodeStatus'
        description: Transcode
      tv_season_id:
        type: string
      tv_show_id:
        description: TV episode fields (nullable; populated only for MediaTypeEpisode
          items)
        type: string
      tv_show_title:
        type: string
      updated_at:
        type: string
      video_codec:
        type: string
      width:
        type: integer
      year:
        type: integer
    type: object
  models.MediaType:
    enum:
    - movie
    - tvshow
    - episode
    - music
    type: string
    x-enum-varnames:
    - MediaTypeMovie
    - MediaTypeTVShow
    - MediaTypeEpisode
    - MediaTypeMusic
  models.RenditionSize:
    properties:
      resolution:
        description: e.g., "360p", "480p", etc.
        type: string
      size:
        description: Total size of all files in the rendition directory in bytes
        type: integer
    type: object
...
  models.TVSeason:
    properties:
      created_at:
        type: string
      id:
        type: string
      overview:
        type: string
      poster_path:
        type: string
      season_number:
        type: integer
      tmdb_id:
        type: integer
      tv_show_id:
        type: string
      updated_at:
        type: string
    type: object
  models.TVShow:
    properties:
      backdrop_path:
        type: string
      cast:
        items:
          $ref: '#/definitions/models.CastMember'
        type: array
      created_at:
        type: string
      director:
        type: string
      extra_posters:
        items:
          type: string
        type: array
      first_air_year:
        type: integer
      id:
        type: string
      library_id:
        type: string
      name:
        type: string
      overview:
        type: string
      poster_path:
        type: string
      tmdb_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.TranscodeSizesInfo:
    properties:
      renditions:
        items:
          $ref: '#/definitions/models.RenditionSize'
        type: array
      total_size:
        type: integer
    type: object
  models.TranscodeStatus:
    enum:
    - none
    - pending
    - processing
    - done
    - failed
    type: string
    x-enum-varnames:
    - TranscodeStatusNone
    - TranscodeStatusPending
    - TranscodeStatusProcessing
    - TranscodeStatusDone
    - TranscodeStatusFailed
  models.TranscodeSubJob:
    properties:
      audio_bitrate_k:
        type: integer
      codec:
        type: string
      created_at:
        type: string
      error_msg:
        type: string
      finished_at:
        type: string
      height:
        type: integer
      id:
        type: string
      job_id:
        type: string
      media_item_id:
        type: string
      profile_id:
        type: string
      profile_name:
        type: string
      progress:
        description: 0-100
        type: number
      started_at:
        type: string
      status:
        $ref: '#/definitions/models.TranscodeStatus'
      type:
        description: '"video" or "subtitles"'
        type: string
      video_bitrate_k:
        type: integer
      width:
        type: integer
      worker_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST & WebSocket API specification for Prism, a self-hosted media server.
  title: Prism Media Server API
  version: "1.0"
paths:
...
  /tv-shows:
    get:
      description: Retrieve a list of all indexed TV shows with optional filtering
        and sorting.
      parameters:
      - description: Library ID
        format: uuid
        in: query
        name: library_id
        type: string
      - description: Sort order
        in: query
        name: sort
        type: string
      - default: 20
        description: Recent items count limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TVShow'
            type: array
        "400":
          description: Invalid library ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List TV Shows
      tags:
      - TV Shows
  /tv-shows/{id}:
    get:
      parameters:
      - description: TV Show ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TVShow'
        "400":
          description: Invalid TV Show ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: TV Show not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get TV Show Details
      tags:
      - TV Shows
  /tv-shows/{id}/seasons:
    get:
      description: Retrieve all seasons belonging to a specific TV show.
      parameters:
      - description: TV Show ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TVSeason'
            type: array
        "400":
          description: Invalid TV Show ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List TV Seasons
      tags:
      - TV Shows
  /tv-shows/{id}/seasons/{number}/episodes:
    get:
      description: Retrieve all episodes (returned as MediaItems) for a specific season
        of a show.
      parameters:
      - description: TV Show ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Season Number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MediaItem'
            type: array
        "400":
          description: Invalid inputs
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Season not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Episodes
      tags:
      - TV Shows
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
  WorkerAuth:
    description: The secret API Key assigned to the transcode worker.
    in: header
    name: X-Worker-API-Key
    type: apiKey
swagger: "2.0"